Refactor: Change photoUrl column type from String to Text in User entity
Detailed description:
The data type of the photoUrl column in the User entity has been updated.
Previously, it was defined as a String with limited length, which restricted the ability to store long URLs.
It has now been changed to a Text type to allow storing long paths or URLs for profile pictures without size restrictions.