Update README.md
feat: Add secondary repositories links in project documentation

The section with links to the project's secondary repositories has been added to the main documentation. The links provide direct access to the backend and frontend repositories of the Gym Membership Management System (GYMETRA). This improves navigation between the different components of the system and allows interested developers to efficiently clone and work with the respective repositories.

Changes made:

The URL for the backend repository was added: GYMETRA_backend

The URL for the frontend repository was added: GYMETRA-V1_frontend

The documentation section was updated to reflect these changes and enhance accessibility.